DAO complete

+++ src/service/StockService.java
@@ -1,6 +1,7 @@
 package service;
 
 import dao.StockDao;
+import dao.WarehouseDao;
 import domain.CheckLog;
 import domain.Stock;
 import domain.Warehouse;
@@ -21,7 +22,13 @@ public class StockService {
         if(stockDao == null) stockDao = StockDao.getInstance();
 
         List<Stock> stockList = stockDao.getAllStockList();
-        if(stockList.isEmpty() || stockList == null) {} //예외처리
+        if(stockList == null){
+            System.out.print("전체 재고 조회 도중 예외가 발생했습니다. ");
+            return null;
+        }else if(stockList.isEmpty()){
+            System.out.print("현재 재고가 존재하지 않습니다. ");
+            return null;
+        }
         return stockList;
     }
 
@@ -34,7 +41,14 @@ public class StockService {
             case 3 -> stockList = stockDao.getSecondaryCategoryStockList(cName);
             case 4 -> stockList = stockDao.getTertiaryCategoryStockList(cName);
         }
-        if(stockList == null || stockList.isEmpty()) {} // 예외처리
+        if(stockList == null){
+            System.out.print("해당 카테고리가 존재하지 않거나 조회 도중 예외가 발생했습니다. ");
+            return null;
+        }else if(stockList.isEmpty()){
+            System.out.print("현재 재고가 존재하지 않습니다. ");
+            return null;
+        }
+
         return stockList;
     }
 
@@ -42,22 +56,33 @@ public class StockService {
         if(stockDao == null) stockDao = StockDao.getInstance();
 
         List<Stock> stockList = stockDao.getProductStockList(pIdx);
-        if(stockList.isEmpty() || stockList == null) {} //예외처리
+        if(stockList == null){
+            System.out.print("바코드 번호에 대한 재고 조회 도중 예외가 발생했습니다. ");
+            return null;
+        }else if(stockList.isEmpty()){
+            System.out.print("현재 재고가 존재하지 않습니다. ");
+            return null;
+        }
         return stockList;
     }
 
-    public CheckLog addCheckLog(){
+    public CheckLog addCheckLog(int wIdx){
         if(stockDao == null) stockDao = StockDao.getInstance();
 
-        int result = stockDao.addCheckLog();
+        int result = stockDao.addCheckLog(wIdx);
         if(result != 1) return null;
 
         CheckLog newCheckLog = stockDao.getNewCheckLog();
         return newCheckLog;
     }
 
-    public int removeCheckLog(int clIdx){
+    public int removeCheckLog(int clIdx, int wIdx, boolean status){
         if(stockDao == null) stockDao = StockDao.getInstance();
+
+        if(status){
+            boolean check = stockDao.checkWarehouseAdminCondition(clIdx, wIdx);
+            if(!check) return -1; //창고관리자가 관리하는 창고가 아님!
+        }
 
         int result = stockDao.removeCheckLog(clIdx);
         return result;
@@ -67,33 +92,62 @@ public class StockService {
         if(stockDao == null) stockDao = StockDao.getInstance();
 
         List<CheckLog> checkLogList = null;
-        //총관리자인 경우
-        checkLogList = stockDao.getCheckLogList(1, 0);
-        //창고관리자인 경우
-        checkLogList = stockDao.getCheckLogList(2, wIdx);
+        if(wIdx == 0){//총관리자인 경우
+            checkLogList = stockDao.getCheckLogList(1, 0);
+        }else{
+            //창고관리자인 경우
+            checkLogList = stockDao.getCheckLogList(2, wIdx);
+        }
+
+        if(checkLogList == null) {
+            System.out.print("재고 실사 로그 조회 중 예외가 발생했습니다."); return null;
+        }else if(checkLogList.isEmpty()) {
+            System.out.print("재고 실사 로그가 기록되지 않았습니다."); return null;
+        }
 
         return checkLogList;
-    }
-
-    public boolean checkWarehouseIsStorage(String wUniqueNum){
-        if(stockDao == null) stockDao = StockDao.getInstance();
-
-        int result = stockDao.checkWarehouseIsStorage(wUniqueNum);
-        if(result == 0) return false;
-        return true;
     }
 
     public List<CheckLog> getSectionCheckLogList(String wUniqueNum, String wsName) {
         if(stockDao == null) stockDao = StockDao.getInstance();
 
+        boolean result = stockDao.checkWsNameExist(wsName);
+        if(!result) {
+            System.out.print("입력하신 창고 섹션은 존재하지 않습니다. ");return null;
+        }
+
+        int result2 = stockDao.checkWarehouseIsStorage(wUniqueNum);
+
+        if(result2 == 0){
+            System.out.print("해당 창고는 마이크로풀필먼트입니다. ");
+            return null;
+        }
+
         List<CheckLog> checkLogList = stockDao.getSectionCheckLoglist(wUniqueNum, wsName);
+
+        if(checkLogList == null) {
+            System.out.print("창고 조회 중 예외가 발생했습니다. "); return null;
+        }else if(checkLogList.isEmpty()){
+            System.out.print("해당 창고 섹션이 존재하지 않습니다. "); return null;
+        }
         return checkLogList;
     }
 
     public List<CheckLog> getWarehouseCheckLogList(String wUniqueNum) {
         if(stockDao == null) stockDao = StockDao.getInstance();
 
+        WarehouseDao temp = WarehouseDao.getInstance();
+        int result = temp.checkWarehouseExist(wUniqueNum);
+        if(result == 0) {
+            System.out.println("해당 창고는 존재하지 않습니다.");
+            return null;
+        }
+
         List<CheckLog> checkLogList = stockDao.getWarehouseCheckLogList(wUniqueNum);
+        if(checkLogList == null) {
+            System.out.print("창고번호별 재고 실사 조회 중 예외가 발생했습니다. ");
+            return null;
+        }
         return checkLogList;
     }
